firebase-push-notifications/requirements.txt
flask
flask-sqlalchemy
request
